notification-service/requirements.txt
aio-pika
sendgrid
python-dotenv
